Refactor: cad_simplification -> simplification
