+++ scripts/cad_simplification/triple_extrusion_simplify.py
@@ -3,7 +3,7 @@ Showcase of the 'TripleExtrusionSimplify' class.
 """
 import time
 
-from volmdlr.cad_simplification import TripleExtrusionSimplify
+from volmdlr.simplification import TripleExtrusionSimplify
 from volmdlr.step import Step
 
 # Load
